update fill and clone commands in line with type hint refactor
